+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/inspections/SpringComponentScanInvalidPackageInspection.kt
@@ -2,17 +2,16 @@ package com.esprito.spring.core.inspections
 
 import com.esprito.inspection.SpringBaseUastLocalInspectionTool
 import com.esprito.spring.core.SpringCoreBundle
-import com.esprito.spring.core.SpringCoreClasses
-import com.esprito.spring.core.SpringCoreClasses.ANNOTATIONS_WITH_PACKAGE_ANT_REFERENCES
 import com.esprito.spring.core.SpringCoreClasses.COMPONENT_SCAN
-import com.esprito.spring.core.util.PsiPackagesSearcher
+import com.esprito.spring.core.SpringCoreClasses.CONFIGURATION_PROPERTIES_SCAN
+import com.esprito.spring.core.search.PsiPackageFqnSearchService
+import com.esprito.spring.core.util.SpringCoreUtil
 import com.esprito.util.EspritoAnnotationUtil
 import com.intellij.codeInspection.InspectionManager
 import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.ElementManipulators
-import com.intellij.psi.search.GlobalSearchScope
 import org.jetbrains.uast.UClass
 
 class SpringComponentScanInvalidPackageInspection : SpringBaseUastLocalInspectionTool() {
@@ -22,22 +21,18 @@ class SpringComponentScanInvalidPackageInspection : SpringBaseUastLocalInspectio
         isOnTheFly: Boolean
     ): Array<ProblemDescriptor> {
         val problems = mutableListOf<ProblemDescriptor>()
+        val packageFqnSearcher = PsiPackageFqnSearchService.getInstance(manager.project)
 
         aClass.uAnnotations
-            .filter { annotation -> ANNOTATIONS_WITH_PACKAGE_ANT_REFERENCES.contains(annotation.qualifiedName) }
             .forEach { annotation ->
                 val attributesName =
                     when (annotation.qualifiedName) {
-                        COMPONENT_SCAN -> setOf("value", "basePackages")
-                        SpringCoreClasses.SPRING_BOOT_APPLICATION -> setOf("scanBasePackages")
-                        else -> emptySet()
+                        COMPONENT_SCAN, CONFIGURATION_PROPERTIES_SCAN -> setOf("value", SpringCoreUtil.BASE_PACKAGES)
+                        else -> setOf(SpringCoreUtil.BASE_PACKAGES, SpringCoreUtil.SCAN_BASE_PACKAGES)
                     }
                 val attributesAsPsiLiteral = EspritoAnnotationUtil.getAttributeValues(annotation, attributesName)
 
-                // if countOfPackagesFound = 0 than we create only 1 problem
-                var countOfPackagesFound = 0
-                val currentProblems = mutableListOf<ProblemDescriptor>()
-                attributesAsPsiLiteral.mapNotNull { it.sourcePsi }.forEach { sourcePsi ->
+                attributesAsPsiLiteral.mapNotNull { it.sourcePsi }.forEach attributes@{ sourcePsi ->
                     val text = ElementManipulators.getValueText(sourcePsi)
                     val range = ElementManipulators.getValueTextRange(sourcePsi)
                     val words = text.split(".")
@@ -46,49 +41,24 @@ class SpringComponentScanInvalidPackageInspection : SpringBaseUastLocalInspectio
                     for (word in words) {
                         path += word
 
-                        val packages = PsiPackagesSearcher.getFilteredPackages(
-                                manager.project,
-                                GlobalSearchScope.allScope(manager.project),
-                                path
-                            )
-                        countOfPackagesFound += packages.size
-
-                        if (packages.isEmpty()) {
-                            var curTextRange = TextRange(
-                                path.length - word.length,
+                        if (!packageFqnSearcher.isPackageExist(path)) {
+                            val curTextRange = TextRange(
+                                0,
                                 path.length
                             ).shiftRight(range.startOffset)
 
-                            //it's for case "org.demo." show problem in "." (not to ")
-                            if (word.isEmpty()) {
-                                curTextRange = curTextRange.shiftLeft(1)
-                            }
-
-                            val problem = manager.createProblemDescriptor(
-                                sourcePsi,
-                                curTextRange,
-                                SpringCoreBundle.message("esprito.spring.inspection.method.componentScan.notFound"),
-                                ProblemHighlightType.ERROR,
-                                true
+                            problems.add(
+                                manager.createProblemDescriptor(
+                                    sourcePsi,
+                                    curTextRange,
+                                    SpringCoreBundle.message("esprito.spring.inspection.method.componentScan.notFound"),
+                                    ProblemHighlightType.ERROR,
+                                    true
+                                )
                             )
-
-                            currentProblems += problem
+                            return@attributes
                         }
-
                         path += "."
-                    }
-
-                    if (countOfPackagesFound == 0) {
-                        val problem = manager.createProblemDescriptor(
-                            sourcePsi,
-                            range,
-                            SpringCoreBundle.message("esprito.spring.inspection.method.componentScan.notFound"),
-                            ProblemHighlightType.ERROR,
-                            true
-                        )
-                        problems += problem
-                    } else {
-                        problems += currentProblems
                     }
                 }
             }