+++ modules/spring-core/testdata/kotlin/inspection/componentScan/src/com/component/scan/TestComponentScan.kt
@@ -39,5 +39,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @ComponentScan(basePackages = "com.?7.?")
 @ComponentScan(["com.2.4", "com.?.?", "com.2.5"])
 @ComponentScan(basePackages = ["com.2.4", "com.?.?", "com.2.5"])
+@ComponentScan("**.3")
+@SpringBootApplication("**.4")
 public class TestComponentScan {
 }