+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### Spring Core/Boot
 
+- Added reference to package from annotation parameters with names `basePackages`, `scanBasePackages`. AntPattern supported
+- Changed behaviour of `Can't find package` inspection. Now it shows inspection till first unknown segment. AntPattern supported
 - Added linemarker from property to library hint
 - Fixed: properties completion for class-reference provider from libraries